requirements.txt
telebot~=0.0.5
pandas~=2.1.1
gspread~=5.11.3
python-dotenv~=1.0.0